refactor: organiza view e arquivos perfil aluno
