+++ src/services/Student/WorkoutsStudentsService.js
@@ -1,4 +1,4 @@
-import api from "./api";
+import api from "../api";
 
 class WorkoutsStudentsService {
   async workoutsByStudentList(studentId) {
